Downloads/biolook/multispecies-train/requirements.txt
torch
torchvision
timm
albumentations
opencv-python
pillow
requests
tqdm
scikit-learn
numpy
pandas
# optional: accelerate distributed training later


